Address comments, update readme
