+++ README.md
@@ -213,6 +213,8 @@ windows_aws_ami             = "<windows ami>"
 no_of_windows_worker_nodes  = <count of Windows node>
 ```
 
+#### NOTE: The sonobuoy test runs outside of the cluster, so if running the test locally, user have to clean up sonobouy manually
+
 ### Test Execution
 
 Tests can be run individually per package:
